test:  add basic unit tests (#1)
test: add Basic Usage test (failing)

The test is based on the example in README.md and verifies that
the result of adding a file to the source directory, and then removing
it again is as expected. (The first test is actually somewhat redundant,
given that the second test includes adding a file as well.)

Some refactoring of storage.js was required to avoid having to call
process.chdir() in tests. Now the fsInterface function takes a basedir
argument, to make all emulated bucket accesses relative to a given
base directory (which is a temporary directory during testing).

Currently, about 1 in 10 runs might fail because of what looks like a
race condition, maybe in jest-resolve. The error looks as follows:

    Cannot find module 'linkedlist' from 'index.js'
      at Resolver.resolveModule (node_modules/jest-resolve/build/index.js:194:17)
      at Object.<anonymous> (node_modules/streaming-cache/index.js:8:18)

If this happens in a Travis CI build or during development, try again.

To complete the integration with Travis CI, a new script "ci" is finally
added to package.json and Node 4 is dropped from the build matrix.

This commit adds new dependencies:

* jest
* jest-fixtures
* eslint-plugin-jest (fixes wrongly reported undefined symbols in tests)
* fs-extra (because Node 6 didn't have fs.copyFile yet)

This commit also enables eslint to recognize ECMAScript 2017
features (async/await).

The async/await feature was introduced in Node 7 (with --harmony)
and is fully supported since Node 8.  However, Cloud Functions still
run with Node 6, so I'm ditching async/await to avoid the whole subject
of code transpilation with Babel.